Move all custom logic for pygments into a dedicated module
This makes it easier to confirm that this is logically separated from
any Sphinx-specific logic, and also makes it easier to test.
